WinMD: split out the table definition (NFC)
This splits out the index representation, table representation, and the
iterator representation to make it easier to work with the types.
